Move module analysis code to Code.Module
